+++ src/Pwa/fileFilter.mjs
@@ -1,3 +1,0 @@
-/**
- * @typedef {(root_file: string) => boolean} fileFilter
- */